* Resolved warning and incorrect type in Xylo-A2 devkit utils * Resolved warnings in XA2 unit test * Ignored json file generated in test bench

+++ tests/tests_default/test_xa2_samna.py
@@ -5,6 +5,7 @@ pytest.importorskip("samna")
 
 def test_from_specification():
     from rockpool.devices.xylo import vA2
+
     # from rockpool.devices.xylo.syns61201 import config_from_specification
     import numpy as np
 
@@ -131,22 +132,24 @@ def test_XyloSamna():
     dt = 1e-3
 
     config, valid, msg = config_from_specification(
-        weights_in=np.random.uniform(-127, 127, size=(Nin, Nhidden, 2)),
-        weights_out=np.random.uniform(-127, 127, size=(Nhidden, Nout)),
-        weights_rec=np.random.uniform(-127, 127, size=(Nhidden, Nhidden, 2)),
-        dash_mem=2 * np.ones(Nhidden),
-        dash_mem_out=3 * np.ones(Nout),
-        dash_syn=4 * np.ones(Nhidden),
-        dash_syn_2=2 * np.ones(Nhidden),
-        dash_syn_out=3 * np.ones(Nout),
-        threshold=128 * np.ones(Nhidden),
-        threshold_out=256 * np.ones(Nout),
+        weights_in=np.random.uniform(-127, 127, size=(Nin, Nhidden, 2)).astype(int),
+        weights_out=np.random.uniform(-127, 127, size=(Nhidden, Nout)).astype(int),
+        weights_rec=np.random.uniform(-127, 127, size=(Nhidden, Nhidden, 2)).astype(
+            int
+        ),
+        dash_mem=2 * np.ones(Nhidden, dtype=int),
+        dash_mem_out=3 * np.ones(Nout, dtype=int),
+        dash_syn=4 * np.ones(Nhidden, dtype=int),
+        dash_syn_2=2 * np.ones(Nhidden, dtype=int),
+        dash_syn_out=3 * np.ones(Nout, dtype=int),
+        threshold=128 * np.ones(Nhidden, dtype=int),
+        threshold_out=256 * np.ones(Nout, dtype=int),
         weight_shift_in=1,
         weight_shift_rec=1,
         weight_shift_out=1,
         aliases=None,
-        bias_hidden=np.ones(Nhidden),
-        bias_out=np.ones(Nout),
+        bias_hidden=np.ones(Nhidden, dtype=int),
+        bias_out=np.ones(Nout, dtype=int),
     )
 
     # - Make a XyloSamna module
